Added Task Manager
For this improvement I added a new function called Task Manager. You can name the task, the priority of it and can select if there is a deadline or not. Then you click on the "Add Task" Button and it is listed. Because we didn't add Login and Backend stuff it is not saved.(Almost Finished, could be improved.)